x86: nSVM: Print out RIP and LBRs from VMCB if LBRV guest test fails
Print out the RIP and LBRs from the VMCB's save are if a LBRV guest test
fails, i.e. exits for a reason other than VMMCALL.  Debugging the guests
is still beyond painful, but with the guest RIP and LBRs, it's at least
possible to piece together things like shutdowns due to unexpected #PF
because the tests neglected to setup the guest stack.

Link: https://lore.kernel.org/r/20230607210959.1577847-7-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>