x86/emulator64: Add macro to test emulation of non-canonical accesses
Add a macro to "try" and check a non-canonical access.  In addition to
de-duplicating the checking logic, this will allow extending the logic to
verify that CR2 isn't incorrectly modified, e.g. on #GP/#SS.

No functional change intended (ignoring the newly added check on the error
code for the JMP case).

Reviewed-by: Mathias Krause <minipli@grsecurity.net>
Tested-by: Mathias Krause <minipli@grsecurity.net>
Link: https://patch.msgid.link/aF1d7rh_vbr8cr7j@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>