x86/svm: Account for numerical rounding errors in TSC scaling test
Fix a rounding error in the nSVM TSC scaling test that causes the test to
run the guest for too short of a duration, which leads to flaky errors on
CPUs with more optimized entry/exit (relative to past CPUs).

On Turing, the test case:

    svm_tsc_scale_run_testcase(50, 0.0001, rdrand());

fails with:

  FAIL: tsc delay (expected: 50, actual: 49)

So, guest_tsc_delay_value is (u64)((50 << 24) * 0.0001), which is
83886.  Note that this is 83886.080000000002 truncated.

If L2 exits after 83886 scaled TSC cycles, the "duration" spent in L2
will be (u64)(83886 / 0.0001) >> 24, which is 49.  To get up to 50, the
CPU has to accumulate an additional (0.080000000002 / 0.0001 =
800.0000000199999) cycles between the two rdtsc() operations bracketing
the svm_vmrun() in L1 .

The test probably passes on other CPUs because emulated VMRUN and
VMEXIT add those 800 cycles.

Instead of truncating ((50 << 24) * 0.0001) and running the guest for too
short of time, calculate the delay as ceil((50 << 24) * 0.0001).

Note, equality of duration and actual_duration is only guaranteed if there
are no significant delays during the measurement, i.e. the test isn't
exactly robust even with the fix.

Reported-by: Srikanth Aithal <sraithal@amd.com>
Signed-off-by: Jim Mattson <jmattson@google.com>
Link: https://patch.msgid.link/CALMp9eQep3H-OtqmLe3O2MsOT-Vx4y0-LordKgN+pkp04VLSWw@mail.gmail.com
[sean: morph Jim's analysis into a changelog]
Signed-off-by: Sean Christopherson <seanjc@google.com>